Added Profile page, edit user info page, and edit password page
