+++ web_interface/account_mgr_app/forms.py
@@ -1,6 +1,7 @@
 from django import forms
-from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.forms import UserCreationForm, UserChangeForm
 from django.contrib.auth.models import User
+from account_mgr_app.models import Profile
 
 
 class SignUpForm(UserCreationForm):
@@ -12,3 +13,13 @@ class SignUpForm(UserCreationForm):
         model = User
         fields = ('username', 'first_name', 'last_name', 'email', 'password1', 'password2', )
         fields = ('username', 'email', 'password1', 'password2', )
+        
+class EditProfileForm(UserChangeForm):
+    
+    class Meta:
+        model = User 
+        fields = {
+            'username',
+            'email',
+            'password'
+            }
