+++ web_interface/account_mgr_app/templates/signup.html
@@ -5,6 +5,8 @@
   <form method="post">
     {% csrf_token %}
     {{ form.as_p }}
+    <p>Don't use SLAC password</p>
+    <p></p>
     <button type="submit">Sign up</button>
   </form>
 {% endblock %}
