+++ web_interface/account_mgr_app/views.py
@@ -1,6 +1,11 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse, HttpResponseRedirect
-from django.contrib.auth.forms import UserCreationForm
+
+from django.contrib.auth.forms import UserCreationForm, UserChangeForm, PasswordChangeForm
+from django.contrib.auth import update_session_auth_hash
+from django.contrib.auth.models import User
+from .forms import EditProfileForm
+
 from django.contrib.auth import login, authenticate
 
 from .forms import SignUpForm
@@ -24,3 +29,39 @@ def signup(request):
         # form = UserCreationForm()
         form = SignUpForm()
     return render(request, 'signup.html', {'form': form})
+    
+    
+    
+    
+def profile(request):
+    args = {'user': request.user}
+    return render(request, 'profile.html', args)
+    
+def edit_profile(request):
+    if request.method == 'POST':
+        form = EditProfileForm(request.POST, instance=request.user)
+        
+        if form.is_valid():
+            form.save()
+            return redirect('/acct/profile')
+            
+    else:
+        form = EditProfileForm(instance=request.user)
+        args = {'form': form}
+        return render(request, 'edit_profile.html', args)
+    
+    
+def change_password(request):
+    if request.method == 'POST':
+        form = PasswordChangeForm(data=request.POST, user=request.user)
+        
+        if form.is_valid():
+            form.save()
+            update_session_auth_hash(request, form.user)
+            return redirect('/acct/profile')
+        else:
+            return redirect('/acct/profile/change_password')
+    else:
+        form = PasswordChangeForm(user=request.user)
+        args = {'form': form}
+        return render(request, 'change_password.html', args)
