+++ web_interface/alert_config_app/views.py
@@ -13,7 +13,7 @@ from django.views.generic.edit import CreateView, UpdateView, DeleteView
 
 from account_mgr_app.models import Profile
 from .models import Alert, Pv, Trigger #PVname #Does this allow you to say "model = Pv....model = Alert..."
-from .forms import configAlert, configTrigger, deleteAlert, subscribeAlert, createPv, EditProfileForm
+from .forms import configAlert, configTrigger, deleteAlert, subscribeAlert, createPv
 
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
@@ -391,36 +391,36 @@ def alert_delete(request,pk=None,*args,**kwargs):
     )
 
 
-def profile(request):
-    args = {'user': request.user}
-    return render(request, 'profile.html', args)
-    
-def edit_profile(request):
-    if request.method == 'POST':
-        form = EditProfileForm(request.POST, instance=request.user)
-        
-        if form.is_valid():
-            form.save()
-            return redirect('/alert/profile')
-            
-    else:
-        form = EditProfileForm(instance=request.user)
-        args = {'form': form}
-        return render(request, 'edit_profile.html', args)
-    
-    
-def change_password(request):
-    if request.method == 'POST':
-        form = PasswordChangeForm(data=request.POST, user=request.user)
-        
-        if form.is_valid():
-            form.save()
-            update_session_auth_hash(request, form.user)
-            return redirect('/alert/profile')
-        else:
-            return redirect('/alert/profile/change_password')
-    else:
-        form = PasswordChangeForm(user=request.user)
-        args = {'form': form}
-        return render(request, 'change_password.html', args)
-        
+#def profile(request):
+#    args = {'user': request.user}
+#    return render(request, 'profile.html', args)
+#    
+#def edit_profile(request):
+#    if request.method == 'POST':
+#        form = EditProfileForm(request.POST, instance=request.user)
+#        
+#        if form.is_valid():
+#            form.save()
+#            return redirect('/alert/profile')
+#            
+#    else:
+#        form = EditProfileForm(instance=request.user)
+#        args = {'form': form}
+#        return render(request, 'edit_profile.html', args)
+#    
+#    
+#def change_password(request):
+#    if request.method == 'POST':
+#        form = PasswordChangeForm(data=request.POST, user=request.user)
+#        
+#        if form.is_valid():
+#            form.save()
+#            update_session_auth_hash(request, form.user)
+#            return redirect('/alert/profile')
+#        else:
+#            return redirect('/alert/profile/change_password')
+#    else:
+#        form = PasswordChangeForm(user=request.user)
+#        args = {'form': form}
+#        return render(request, 'change_password.html', args)
+#
