+++ web_interface/templates/registration/login.html
@@ -8,4 +8,6 @@
     {{ form.as_p }}
     <button type="submit">Login</button>
   </form>
+  <br></br>
+  <a href="{% url 'reset_password' %}">Forgot your password?</a>
   {% endblock %}
