+++ web_interface/templates/registration/password_reset_confirm.html
@@ -1,0 +1,42 @@
+{% extends "base.html" %}
+{% load i18n static %}
+
+{% block extrastyle %}{{ block.super }}<link rel="stylesheet" type="text/css" href="{% static "admin/css/forms.css" %}" />{% endblock %}
+{% block breadcrumbs %}
+<div class="breadcrumbs">
+<a href="{% url 'admin:index' %}">{% trans 'Home' %}</a>
+&rsaquo; {% trans 'Password reset confirmation' %}
+</div>
+{% endblock %}
+
+{% block title %}{{ title }}{% endblock %}
+{% block content_title %}<h1>{{ title }}</h1>{% endblock %}
+{% block content %}
+
+{% if validlink %}
+
+<p>{% trans "Please enter your new password twice so we can verify you typed it in correctly." %}</p>
+
+<form method="post">{% csrf_token %}
+<fieldset class="module aligned">
+	<div class="form-row field-password1">
+		{{ form.new_password1.errors }}
+		<label for="id_new_password1">{% trans 'New password:' %}</label>
+		{{ form.new_password1 }}
+	</div>
+	<div class="form-row field-password2">
+		{{ form.new_password2.errors }}
+		<label for="id_new_password2">{% trans 'Confirm password:' %}</label>
+		{{ form.new_password2 }}
+	</div>
+	<input type="submit" value="{% trans 'Change my password' %}" />
+</fieldset>
+</form>
+
+{% else %}
+
+<p>{% trans "The password reset link was invalid, possibly because it has already been used.  Please request a new password reset." %}</p>
+
+{% endif %}
+
+{% endblock %}
