+++ web_interface/templates/registration/password_reset_form.html
@@ -1,0 +1,29 @@
+{% extends "base.html" %}
+{% load i18n static %}
+
+{% block extrastyle %}{{ block.super }}<link rel="stylesheet" type="text/css" href="{% static "admin/css/forms.css" %}" />{% endblock %}
+{% block breadcrumbs %}
+<div class="breadcrumbs">
+<a href="{% url 'admin:index' %}">{% trans 'Home' %}</a>
+&rsaquo; {% trans 'Password reset' %}
+</div>
+{% endblock %}
+
+{% block title %}{{ title }}{% endblock %}
+{% block content_title %}<h1>{{ title }}</h1>{% endblock %}
+{% block content %}
+
+<p>{% trans "Forgotten your password? Enter your email address below, and we'll email instructions for setting a new one." %}</p>
+
+<form method="post">{% csrf_token %}
+<fieldset class="module aligned">
+	<div class="form-row field-email">
+		{{ form.email.errors }}
+		<label for="id_email">{% trans 'Email address:' %}</label>
+		{{ form.email }}
+	</div>
+	<input type="submit" value="{% trans 'Reset my password' %}" />
+</fieldset>
+</form>
+
+{% endblock %}
